Renamed Engine.hpp and Engine.cpp to PhysicsWorld
